Moving hegel::hegel to the cpp file so that HEGEL_DEFAULT_PATH isn't overwritten at runtime
